// DIMS Client/app/model/taskItems.model.ts
export class TaskItems {
  workflowWorkItemID: string;
  workflowSender: string;
  workflowStepNo: number;
  workflowItemReceivedOn: number;
  workflowItemDeadline: string;
  workflowItemStatus: string;
  workflowItemSenderDepartment: number;
  workflowItemReceiverDepartment: number;
  workflowItemSenderDiv: number;
  workflowItemReceiverDiv: number;
  workflowItemSubject: string;
  workflowItemPriority: string;
  workflowSenderName: string;
  workflowWorkItemType: string;
  workflowItemRootSender: string;
  is_checked: boolean;
  totalCount: any;
  isOverdue: string;
}